feat: updated page layout for the Dashboard, My Orders and Change Password
