+++ client/src/App.jsx
@@ -17,6 +17,8 @@ import AccountSettingsPage from "./pages/AccountSettingsPage";
 import AdminManagementPage from "./pages/AdminManagementPage";
 import AdminAccountsPage from "./pages/AdminAccountsPage";
 import DashboardPage from "./pages/DashboardPage";
+import ChangePasswordPage from "./pages/ChangePasswordPage";
+import DashboardLayout from "./layouts/DashboardLayout";
 
 import { clearExpiredSession, getStoredUser, hasValidSession } from "./utils/authSession";
 
@@ -106,22 +108,16 @@ function App() {
       />
 
       <Route
-        path="/account/dashboard"
         element={
           <ProtectedRoute>
-            <DashboardPage />
+            <DashboardLayout />
           </ProtectedRoute>
         }
-      />
-
-      <Route
-        path="/account/orders"
-        element={
-          <ProtectedRoute>
-            <CustomerOrdersPage />
-          </ProtectedRoute>
-        }
-      />
+      >
+        <Route path="/account/dashboard" element={<DashboardPage />} />
+        <Route path="/account/orders" element={<CustomerOrdersPage />} />
+        <Route path="/account/change-password" element={<ChangePasswordPage />} />
+      </Route>
 
       <Route
         path="/provider/orders/:id"
