+++ server/prisma/schema.prisma
@@ -42,14 +42,17 @@ model User {
 }
 
 model Profile {
-  id          String   @id @default(cuid())
-  userId      String   @unique
-  displayName String?
-  bio         String?
+  id              String    @id @default(cuid())
+  userId          String    @unique
+  displayName     String?
+  bio             String?
   profileImageUrl String?
-  createdAt   DateTime @default(now())
-  updatedAt   DateTime @updatedAt
-  user        User     @relation(fields: [userId], references: [id], onDelete: Cascade)
+  discord         String?
+  country         String?
+  birthday        DateTime?
+  createdAt       DateTime  @default(now())
+  updatedAt       DateTime  @updatedAt
+  user            User      @relation(fields: [userId], references: [id], onDelete: Cascade)
 }
 
 model Service {
